+++ Samples/Showcase/Editor/EditorUitkxDeferredEffectDemoWindow.cs
@@ -3,7 +3,7 @@ using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;
 using ReactiveUITK.EditorSupport;
-using ReactiveUITK.Samples.UITKXComponents;
+using ReactiveUITK.Samples.Components.DeferredEffectDemoFunc;
 
 namespace ReactiveUITK.Samples.UITKX.Editor
 {